==> pandas/1-series.ipynb <==
import pandas as pd

dados = [10, 20, 30, 40, 50]
serie = pd.Series(dados)
serie

type(serie)

#1-Criando uma Series com dados de vendas mensais
vendas_mensais = pd.Series(
    [2000, 2500, 3000, 2800, 3500, 4000, 4500, 5000, 4800, 5200, 6000, 7000],
    index=['Jan', 'Fev', 'Mar', 'Abr', 'Mai', 'Jun', 'Jul', 'Ago', 'Set', 'Out', 'Nov', 'Dez']
)

vendas_mensais

#2-Calculando e exibir a média de vendas
media_vendas = vendas_mensais.mean()
media_vendas

#3-Identificando o mês com maior venda
mes_maior_venda = vendas_mensais.idxmax()
mes_maior_venda

#4-Exibir as vendas acima da média
vendas_acima_media = vendas_mensais[vendas_mensais > media_vendas]
vendas_acima_media
