==> pandas/7-cria_planilha.ipynb <==
import pandas as pd

dados = {
    'Times': ['Flamengo', 'Palmeiras', 'São Paulo', 'Santos', 'Corinthians'],
    'Partidas_Jogadas': [10, 10, 10, 10, 10],
    'Vitorias': [7, 5, 6, 4, 3],
    'Empates': [2, 3, 1, 3, 2],
    'Derrotas': [1, 2, 3, 3, 5],
    'Gols_Pro': [22, 18, 16, 14, 10],
    'Gols_Contra': [10, 12, 15, 16, 20]  
}

df = pd.DataFrame(dados)
df

df['Saldo_Gols'] = df['Gols_Pro'] - df['Gols_Contra']
df

df['Pontos'] = df['Vitorias'] * 3 + df['Empates'] * 1
df

df = df.sort_values(by='Pontos', ascending=False)
df

df.to_excel("dados/campeonato_futebol.xlsx", index=False)
